common/requirements.txt
python-dotenv
openai
google-cloud-aiplatform
anthropic
groq
stability_sdk
pika
